Moved stand-alone examples out of the ROS folder.
