+++ keyence_library/src/change_program.cpp
@@ -7,19 +7,20 @@
 
 int main(int argc, char** argv)
 {
+  using keyence::command::ChangeProgram;
+  using keyence::Client;
+
+  if (argc != 4)
+  {
+    std::cerr << "Usage: ./keyence_change_program <host> <port> <program_no>\n";
+    return 1;
+  }
+
   try
   {
-    using keyence::command::ChangeProgram;
-    using keyence::Client;
-
-    if (argc != 4)
-    {
-      std::cerr << "Usage: ./keyence_change_program <host> <port> <program_no>\n";
-    }
-
     keyence::TcpClient keyence (argv[1], argv[2]);
 
-    ChangeProgram::Request req(std::atoi(argv[3]));
+    ChangeProgram::Request req (std::atoi(argv[3]));
 
     Client::Response<ChangeProgram::Request> resp = keyence.sendReceive(req);
 
@@ -28,7 +29,8 @@ int main(int argc, char** argv)
   catch (const keyence::KeyenceException& exc)
   {
     std::cerr << exc.what();
+    return -1;
   }
   
-return 0;
+  return 0;
 }
