+++ package.xml
@@ -21,5 +21,6 @@
   
   <run_depend>roscpp</run_depend>
   <run_depend>pcl_ros</run_depend>
+  <run_depend>message_runtime</run_depend>
 
 </package>
